+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to Semantic Versioning.
 
 ## [Unreleased]
+### Fixed
+- Avoid memory leak by unsubscribing from `ReceivedMessageFrame` event.
 
 ## [1.0.0-alpha3] - 2018-10-11
 ### Added